pyproject.toml
[tool.poetry]
package-mode = false
name = "test-checker"
version = "0.1.0"
description = ""
authors = ["Marko Yaminsky <marko.yaminskiy03@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pyjwt = "^2.10.1"
fastapi = "^0.115.6"
alembic = "^1.14.1"
sqlalchemy = "^2.0.37"
pydantic = "^2.10.5"
pydantic-settings = "^2.7.1"
psycopg = {extras = ["binary", "pool"], version = "^3.2.4"}
uvicorn = "^0.34.0"
passlib = "1.7.4"
bcrypt = "4.0.1"
pre-commit = "^4.1.0"
boto3 = "^1.36.6"
python-multipart = "^0.0.20"
greenlet = "^3.1.1"
celery = "^5.4.0"
redis = "^5.2.1"
openai = "^1.61.1"
reportlab = "^4.3.0"
fpdf2 = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
